[csl] take into account radio header for aiming at the CSL phase
This commit introduces a shift of -192 us in the target time for CSL
transmission/reception operations to take into account that the radios
should allow for SHR and PHR to be transmitted/received pior to the
timestamp marked by the CSL phase.

This commit also takes into account local radio uncertainty for
calculating the required CSL receive window duration.